+++ src/mounters/fluent-functions.ts
@@ -1,3 +1,4 @@
+import { distinctByAsync } from './../async/distinct-by-async';
 import {
   aggregate,
   any,
@@ -32,6 +33,7 @@ import {
   sort,
   sortBy,
   distinct,
+  distinctBy,
   group,
   last,
   all,
@@ -118,6 +120,7 @@ export const iterableFuncs = {
   sort,
   sortBy,
   distinct,
+  distinctBy,
   execute,
   combine,
   combineJoin,
@@ -138,6 +141,7 @@ export const iterableAsyncFuncs = {
   flatMerge,
   flatMapAsync: flattenAsync,
   distinctAsync,
+  distinctByAsync,
   executeAsync,
   toAsync,
   combineAsync,