+++ src/sync/to-object-chain.ts
@@ -1,14 +1,14 @@
 import { toObjectChainRecipe } from '../recipes';
 import { basic } from './basic-ingredients';
+import { flatten } from './flatten';
 import { group } from './group';
 import { reduce } from './reduce';
 import { toObject } from './to-object';
-import { unwind } from './unwind';
 
 export const toObjectChain = toObjectChainRecipe({
   ...basic,
   group,
   toObject,
-  unwind,
+  flatten,
   reduce,
 });